add new fritzdect api
