+++ meter/fritzdect_new.go
@@ -1,0 +1,28 @@
+package meter
+
+import (
+	"github.com/evcc-io/evcc/api"
+	"github.com/evcc-io/evcc/meter/fritzdect_new"
+	"github.com/evcc-io/evcc/util"
+)
+
+// AVM FritzBox AHA interface specifications:
+// https://avm.de/fileadmin/user_upload/Global/Service/Schnittstellen/AHA-HTTP-Interface.pdf
+
+func init() {
+	registry.Add("fritzdect_new", NewFritzDECTFromConfig2)
+}
+
+// NewFritzDECTFromConfig creates a fritzdect meter from generic config
+func NewFritzDECTFromConfig2(other map[string]any) (api.Meter, error) {
+	var cc fritzdect_new.Settings
+	if err := util.DecodeOther(other, &cc); err != nil {
+		return nil, err
+	}
+
+	if cc.User == "" || cc.Password == "" {
+		return nil, api.ErrMissingCredentials
+	}
+
+	return fritzdect_new.NewConnection(cc.URI, cc.AIN, cc.User, cc.Password)
+}
